Fix build against 6.0
Fixes: QTBUG-89381

+++ src/httpserver/qhttpserverrequest.cpp
@@ -125,7 +125,7 @@ bool QHttpServerRequestPrivate::parse(QIODevice *socket)
     return true;
 }
 
-uint QHttpServerRequestPrivate::headerHash(const QByteArray &key) const
+size_t QHttpServerRequestPrivate::headerHash(const QByteArray &key) const
 {
     return qHash(key.toLower(), headersSeed);
 }
